feat(registration): add lots of styling
