+++ frontend/src/components/Footer/Footer.tsx
@@ -1,6 +1,6 @@
 export default function Footer() {
   return (
-    <nav className="mt-auto bg-primary">
+    <nav className="mt-auto bg-primary bottom-0">
       <div className="max-w-screen items-center justify-between mx-auto p-4">
         <div className="w-full" id="navbar-default">
           <ul className="font-medium flex items-center h-5 md:p-0 mt-4 md:flex-row md:space-x-8 md:mt-0">
