+++ frontend/src/components/Layout/Layout.tsx
@@ -6,7 +6,7 @@ export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <div
       id="layout"
-      className="flex min-h-screen flex-col bg-background text-body-text"
+      className="flex h-screen flex-col bg-background text-body-text"
     >
       <Navbar />
       {children}
