+++ frontend/src/utilities/commonPrinters.ts
@@ -1,0 +1,49 @@
+import { Printer } from "../main.types";
+
+export const commonPrinters: Printer[] = [
+    {
+        name: "Bambu",
+        supportedFilament: ["PLA", "ABS"],
+        dimensions: {
+            height: 200,
+            width: 200,
+            depth: 200,
+        },
+    },
+    {
+        name: "Creality",
+        supportedFilament: ["PLA", "ABS"],
+        dimensions: {
+            height: 200,
+            width: 200,
+            depth: 200,
+        },
+    },
+    {
+        name: "Sovol",
+        supportedFilament: ["PLA", "ABS"],
+        dimensions: {
+            height: 200,
+            width: 200,
+            depth: 200,
+        },
+    },
+    {
+        name: "Elegoo",
+        supportedFilament: ["PLA", "ABS"],
+        dimensions: {
+            height: 200,
+            width: 200,
+            depth: 200,
+        },
+    },
+    {
+        name: "Prusa",
+        supportedFilament: ["PLA", "ABS"],
+        dimensions: {
+            height: 200,
+            width: 200,
+            depth: 200,
+        },
+    },
+];
